refactor(git): split the worktree module into focused submodules
